requirements.txt
annotated-types==0.7.0
anyio==4.8.0
anthropic==0.21.1
browserbase==1.2.0
certifi==2025.1.31
charset-normalizer==3.4.1
distro==1.9.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jiter==0.8.2
pillow==11.1.0
playwright==1.50.0
pyautogui==0.9.54
pywin32==306; platform_system=="Windows"
pydantic==2.10.6
pydantic_core==2.27.2
pyee==12.1.1
python-dotenv==1.0.1
requests==2.32.3
scrapybara>=2.3.6
sniffio==1.3.1
typing_extensions==4.12.2
urllib3==2.3.0
